requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
faker>=24.0.0
rich>=13.0.0
python-dateutil>=2.9.0